Completed App
This is not completed to my actual desires, but for the purpose of this
class, this app is complete. I will go back and revise and delve deeper
into building this app to allow the Connect feature as well as deeper
Tasks and the function to add/delete activities as desired.